+++ zarrs/src/array.rs
@@ -899,8 +899,29 @@ impl<TStorage: ?Sized> Array<TStorage> {
         }
     }
 
-    /// Reject the array if it contains additional fields with `"must_understand": true`.
+    /// Reject the array if it contains unsupported extensions or additional fields with `"must_understand": true`.
     fn validate_metadata(metadata: &ArrayMetadata) -> Result<(), ArrayCreateError> {
+        match &metadata {
+            ArrayMetadata::V2(_) => {}
+            ArrayMetadata::V3(metadata) => {
+                for extension in &metadata.extensions {
+                    if extension.must_understand() {
+                        return Err(ArrayCreateError::UnsupportedAdditionalFieldError(
+                            UnsupportedAdditionalFieldError::new(
+                                extension.name().to_string(),
+                                extension
+                                    .configuration()
+                                    .map(|configuration| {
+                                        serde_json::Value::Object(configuration.clone().into())
+                                    })
+                                    .unwrap_or_default(),
+                            ),
+                        ));
+                    }
+                }
+            }
+        }
+
         let additional_fields = match &metadata {
             ArrayMetadata::V2(metadata) => &metadata.additional_fields,
             ArrayMetadata::V3(metadata) => &metadata.additional_fields,